Refactor FourUpQuiz to be QuizDeck
This paves the way to allow for other types of questions in the deck.
